style: move focus funcs to hook
